requirements.txt
Pillow==10.0.1
beautifulsoup4==4.11.1
numpy==1.23.2
transformers==4.21.
sentencepiece==0.1.99
timm==0.5.4
--extra-index-url https://download.pytorch.org/whl/cu116
torch==1.12.1+cu116
torchaudio==0.12.1+cu116
torchvision==0.13.1+cu116
paddleocr==2.7.0.3